# server/docs/swagger.yaml
basePath: /
definitions:
...
  model.Rating:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      numeric_value:
        type: integer
      slug:
        type: string
      space_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.Space:
    properties:
      contact_info:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      organisation_id:
        type: integer
      site_address:
        type: string
      site_title:
        type: string
      slug:
        type: string
      social_media_urls:
        type: string
      tag_line:
        type: string
      updated_at:
        type: string
      verification_codes:
        type: string
    type: object
...
  rating.paging:
    properties:
      nodes:
        items:
          $ref: '#/definitions/model.Rating'
        type: array
      total:
        type: integer
    type: object
  rating.rating:
    properties:
      description:
        type: string
      name:
        type: string
      numeric_value:
        type: integer
      slug:
        type: string
    required:
    - name
    - numeric_value
    type: object
  space.space:
    properties:
      contact_info:
        type: string
      description:
        type: string
      name:
        type: string
      organisation_id:
        type: integer
      site_address:
        type: string
      site_title:
        type: string
      slug:
        type: string
      social_media_urls:
        type: string
      tag_line:
        type: string
      verification_codes:
        type: string
    required:
    - name
    - organisation_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: VidCheck server API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: VidCheck API
  version: "1.0"
paths:
...
  /ratings:
    get:
      description: Get all ratings
      operationId: get-all-ratings
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      - description: limit per page
        in: query
        name: limit
        type: string
      - description: page number
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rating.paging'
      summary: Show all ratings
      tags:
      - Rating
    post:
      description: Create rating
      operationId: add-rating
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      - description: Rating Object
        in: body
        name: Rating
        required: true
        schema:
          $ref: '#/definitions/rating.rating'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Rating'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Create rating
      tags:
      - Rating
  /ratings/{rating_id}:
    delete:
      description: Delete rating by ID
      operationId: delete-rating-by-id
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      - description: Rating ID
        in: path
        name: rating_id
        required: true
        type: string
      responses:
        "200": {}
      summary: Delete a rating
      tags:
      - Rating
    get:
      description: Get rating by ID
      operationId: get-rating-by-id
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      - description: Rating ID
        in: path
        name: rating_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rating'
      summary: Show a rating by id
      tags:
      - Rating
    put:
      description: Update rating by ID
      operationId: update-rating-by-id
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      - description: Rating ID
        in: path
        name: rating_id
        required: true
        type: string
      - description: Rating
        in: body
        name: Rating
        schema:
          $ref: '#/definitions/rating.rating'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rating'
      summary: Update a rating by id
      tags:
      - Rating
  /ratings/default:
    post:
      description: Create Default Ratings
      operationId: add-default-ratings
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/model.Rating'
            type: array
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Create Default Ratings
      tags:
      - Rating
  /spaces:
    get:
      description: Get all spaces
      operationId: get-all-spaces
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/model.Space'
              type: array
            type: array
      summary: Show all spaces
      tags:
      - Space
    post:
      description: Create space
      operationId: add-space
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space Object
        in: body
        name: Space
        required: true
        schema:
          $ref: '#/definitions/space.space'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Space'
      summary: Create space
      tags:
      - Space
  /spaces/{space_id}:
    delete:
      description: Delete space
      operationId: delete-space
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: path
        name: space_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      summary: Delete space
      tags:
      - Space
    put:
      description: Update space
      operationId: update-space
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: path
        name: space_id
        required: true
        type: string
      - description: Space Object
        in: body
        name: Space
        required: true
        schema:
          $ref: '#/definitions/space.space'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Space'
      summary: Update space
      tags:
      - Space
...
swagger: "2.0"